refactor: Serial PO line info
Seperated out duplicate code from SerialPOLine and POLineForm into a seperate component
