+++ src/components/SerialPOLineInfo/index.js
@@ -1,0 +1,1 @@
+export { default } from './SerialPOLineInfo';
